+++ OptimizelySDK.Tests/OptimizelyTest.cs
@@ -28,6 +28,7 @@ using OptimizelySDK.Tests.UtilsTests;
 using OptimizelySDK.Bucketing;
 using OptimizelySDK.Notifications;
 using OptimizelySDK.Tests.NotificationTests;
+using OptimizelySDK.Utils;
 
 namespace OptimizelySDK.Tests
 {
@@ -38,6 +39,7 @@ namespace OptimizelySDK.Tests
         private ProjectConfig Config;
         private Mock<EventBuilder> EventBuilderMock;
         private Mock<IErrorHandler> ErrorHandlerMock;
+        private Mock<IEventDispatcher> EventDispatcherMock;
         private Optimizely Optimizely;
         private IEventDispatcher EventDispatcher;
         private const string TestUserId = "testUserId";
@@ -73,19 +75,19 @@ namespace OptimizelySDK.Tests
                 logger: LoggerMock.Object,
                 errorHandler: new NoOpErrorHandler());
 
-            EventDispatcher = new ValidEventDispatcher();
-            Optimizely = new Optimizely(TestData.Datafile, EventDispatcher, LoggerMock.Object, ErrorHandlerMock.Object);
+            EventDispatcherMock = new Mock<IEventDispatcher>();
+            Optimizely = new Optimizely(TestData.Datafile, EventDispatcherMock.Object, LoggerMock.Object, ErrorHandlerMock.Object);
 
             Helper = new OptimizelyHelper
             {
                 Datafile = TestData.Datafile,
-                EventDispatcher = EventDispatcher,
+                EventDispatcher = EventDispatcherMock.Object,
                 Logger = LoggerMock.Object,
                 ErrorHandler = ErrorHandlerMock.Object,
                 SkipJsonValidation = false,
             };
 
-            OptimizelyMock = new Mock<Optimizely>(TestData.Datafile, EventDispatcher, LoggerMock.Object, ErrorHandlerMock.Object, null, false)
+            OptimizelyMock = new Mock<Optimizely>(TestData.Datafile, EventDispatcherMock.Object, LoggerMock.Object, ErrorHandlerMock.Object, null, false)
             {
                 CallBase = true
             };
@@ -515,8 +517,7 @@ namespace OptimizelySDK.Tests
 
             Optimizely.Track("purchase", TestUserId, attributes);
 
-            //LoggerMock.Verify(l => l.Log(LogLevel.ERROR, "Provided attributes are in an invalid format."), Times.Once);
-            ErrorHandlerMock.Verify(e => e.HandleError(It.IsAny<InvalidAttributeException>()), Times.Once);
+            LoggerMock.Verify(l => l.Log(LogLevel.ERROR, @"Attribute key ""abc"" is not in datafile."), Times.Once);
         }
 
         [Test]
@@ -1111,7 +1112,7 @@ namespace OptimizelySDK.Tests
                { "device_type", "iPhone" },
                { "company", "Optimizely" },
                { "location", "San Francisco" },
-               { DecisionService.RESERVED_ATTRIBUTE_KEY_BUCKETING_ID, testBucketingIdVariation }
+               { ControlAttributes.BUCKETING_ID_ATTRIBUTE, testBucketingIdVariation }
             };
 
             // confirm that a valid variation is bucketed without the bucketing ID
@@ -1537,6 +1538,34 @@ namespace OptimizelySDK.Tests
 
             LoggerMock.Verify(l => l.Log(LogLevel.INFO,
                 $@"Feature flag ""{featureKey}"" is enabled for user ""{TestUserId}""."));
+        }
+
+        // Should return false and send an impression event when feature is enabled for the user 
+        // and user is being experimented.
+        [Test]
+        public void TestIsFeatureEnabledGivenFeatureFlagIsNotEnabledAndUserIsBeingExperimented()
+        {
+            var featureKey = "double_single_variable_feature";
+            var experiment = Config.GetExperimentFromKey("test_experiment_double_feature");
+            var variation = Config.GetVariationFromKey("test_experiment_double_feature", "variation");
+            var featureFlag = Config.GetFeatureFlagFromKey(featureKey);
+            var decision = new FeatureDecision(experiment, variation, FeatureDecision.DECISION_SOURCE_EXPERIMENT);
+
+            DecisionServiceMock.Setup(ds => ds.GetVariationForFeature(featureFlag, TestUserId, null)).Returns(decision);
+
+            var optly = Helper.CreatePrivateOptimizely();
+            optly.SetFieldOrProperty("DecisionService", DecisionServiceMock.Object);
+
+            bool result = (bool)optly.Invoke("IsFeatureEnabled", featureKey, TestUserId, null);
+            Assert.False(result);
+
+            // SendImpressionEvent() gets called.
+            LoggerMock.Verify(l => l.Log(LogLevel.INFO,
+                $@"The user ""{TestUserId}"" is not being experimented on feature ""{featureKey}""."), Times.Never);
+
+            LoggerMock.Verify(l => l.Log(LogLevel.INFO,
+                $@"Feature flag ""{featureKey}"" is not enabled for user ""{TestUserId}""."));
+            EventDispatcherMock.Verify(dispatcher => dispatcher.DispatchEvent(It.IsAny<LogEvent>()));
         }
 
         // Verify that IsFeatureEnabled returns true if a variation does not get found in the feature